limpeza da classe Controller

+++ src/main/java/br/com/goldsgym/cadastro/controller/service/AlunoService.java
@@ -7,8 +7,10 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.com.goldsgym.cadastro.controller.dto.AlunoDto;
 import br.com.goldsgym.cadastro.modelo.Aluno;
@@ -18,11 +20,12 @@ import br.com.goldsgym.cadastro.repository.AlunoRepository;
 public class AlunoService {
 
 	final AlunoRepository alunoRepository;
-	
+	//private AlunoService alunoService;
+
 	public AlunoService(AlunoRepository alunoRepository) {
 		this.alunoRepository = alunoRepository;
 	}
-	
+
 //  Método para cadastrar um aluno(POST)
 	@Transactional
 	public Aluno save(Aluno alunoDto) {
@@ -38,18 +41,31 @@ public class AlunoService {
 
 //	Método para buscar um determinado aluno pelo id(GET(ID))
 	public Optional<Aluno> findById(Long id) {
+		boolean isAlunoPresent = alunoRepository.isAlunoPresent(id);
+		if (!isAlunoPresent) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Aluno não encontrado");
+		}
 		return alunoRepository.findById(id);
 	}
 
 //  Método para deletar um aluno(DELETE)		
 	@Transactional
 	public void deleteById(Long id) {
+		boolean isAlunoPresent = alunoRepository.isAlunoPresent(id);
+		if(!isAlunoPresent) {
+			//return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Aluno não cadastrado");
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Aluno não cadastrado");
+		}	
 		alunoRepository.deleteById(id);
 	}
-	
+
 //  Método para alterar/atualizar o cadastro de um aluno(PUT)		
-	public Aluno setDtoToObject(@PathVariable Long id, @Valid AlunoDto alunoDto) { 
+	public Aluno setDtoToObject(@PathVariable Long id, @Valid AlunoDto alunoDto) {
 		Optional<Aluno> alunoOptional = alunoRepository.findById(id);
+		boolean isAlunoPresent = alunoRepository.isAlunoPresent(id);
+		if(!isAlunoPresent) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Aluno não cadastrado");
+		}
 		Aluno alunoAtualizado = new Aluno();
 		BeanUtils.copyProperties(alunoDto, alunoAtualizado);
 		alunoAtualizado.setId(alunoOptional.get().getId());
@@ -58,14 +74,14 @@ public class AlunoService {
 
 	public boolean isAlunoPresent(Long id) {
 		Optional<Aluno> alunoOptional = findById(id);
-		if(!alunoOptional.isPresent()) {
+		if (!alunoOptional.isPresent()) {
 			return false;
 		}
 		return true;
 	}
-	
+
 	public Aluno get(Long id) {
 		return alunoRepository.getOne(id);
 	}
-	
+
 }
